Adjusting the code to be more readable
